+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 > Released 29 Jul 2025
 
 * Fixed: `Unprotect-CFileToken` fails when `OutputPath` is a relative path.
+* Fixed: `Unprotect-CFileToken` shouldn't require `Force` when `OutputPath` already exists but is empty.
 
 ## 3.5.0